Moved token generation methods into its own module
